+++ src/game/Object/Player.cpp
@@ -5295,53 +5295,6 @@ bool Player::IsAffectedBySpellmod(SpellEntry const* spellInfo, SpellModifier* mo
 }
 
 /**
- * @brief Applies or removes a spell modifier and notifies the client.
- *
- * @param mod The modifier to add or remove.
- * @param apply True to apply the modifier; false to remove it.
- */
-void Player::AddSpellMod(SpellModifier* mod, bool apply)
-{
-    OpcodesList opcode = (mod->type == SPELLMOD_FLAT) ? SMSG_SET_FLAT_SPELL_MODIFIER : SMSG_SET_PCT_SPELL_MODIFIER;
-
-    for (int eff = 0; eff < 64; ++eff)
-    {
-        uint64 _mask = uint64(1) << eff;
-        if (mod->mask.IsFitToFamilyMask(_mask))
-        {
-            int32 val = 0;
-            for (SpellModList::const_iterator itr = m_spellMods[mod->op].begin(); itr != m_spellMods[mod->op].end(); ++itr)
-            {
-                if ((*itr)->type == mod->type && ((*itr)->mask.IsFitToFamilyMask(_mask)))
-                {
-                    val += (*itr)->value;
-                }
-            }
-            val += apply ? mod->value : -(mod->value);
-            WorldPacket data(opcode, (1 + 1 + 4));
-            data << uint8(eff);
-            data << uint8(mod->op);
-            data << int32(val);
-            SendDirectMessage(&data);
-        }
-    }
-
-    if (apply)
-    {
-        m_spellMods[mod->op].push_back(mod);
-    }
-    else
-    {
-        if (mod->charges == -1)
-        {
-            --m_SpellModRemoveCount;
-        }
-        m_spellMods[mod->op].remove(mod);
-        delete mod;
-    }
-}
-
-/**
  * @brief Finds a spell modifier by operation and owning spell.
  *
  * @param op The modifier operation bucket.
@@ -6371,99 +6324,6 @@ void Player::InitPrimaryProfessions()
     uint32 maxProfs = GetSession()->GetSecurity() < AccountTypes(sWorld.getConfig(CONFIG_UINT32_TRADE_SKILL_GMIGNORE_MAX_PRIMARY_COUNT))
                       ? sWorld.getConfig(CONFIG_UINT32_MAX_PRIMARY_TRADE_SKILL) : 10;
     SetFreePrimaryProfessions(maxProfs);
-}
-
-/**
- * @brief Sends the current combo point target and value to the client fields.
- */
-void Player::SendComboPoints()
-{
-    Unit* combotarget = sObjectAccessor.GetUnit(*this, m_comboTargetGuid);
-    if (combotarget)
-    {
-        WorldPacket data(SMSG_UPDATE_COMBO_POINTS, combotarget->GetPackGUID().size() + 1);
-        data << combotarget->GetPackGUID();
-        data << uint8(m_comboPoints);
-        GetSession()->SendPacket(&data);
-    }
-    /*else
-    {
-        // can be NULL, and then points=0. Use unknown; to reset points of some sort?
-        data << PackedGuid();
-        data << uint8(0);
-        GetSession()->SendPacket(&data);
-    }*/
-}
-
-/**
- * @brief Adds combo points on a target and updates combo point ownership.
- *
- * @param target The unit receiving combo points.
- * @param count The number of combo points to add.
- */
-void Player::AddComboPoints(Unit* target, int8 count)
-{
-    if (!count)
-    {
-        return;
-    }
-
-    // without combo points lost (duration checked in aura)
-    RemoveSpellsCausingAura(SPELL_AURA_RETAIN_COMBO_POINTS);
-
-    if (target->GetObjectGuid() == m_comboTargetGuid)
-    {
-        m_comboPoints += count;
-    }
-    else
-    {
-        if (m_comboTargetGuid)
-            if (Unit* target2 = sObjectAccessor.GetUnit(*this, m_comboTargetGuid))
-            {
-                target2->RemoveComboPointHolder(GetGUIDLow());
-            }
-
-        m_comboTargetGuid = target->GetObjectGuid();
-        m_comboPoints = count;
-
-        target->AddComboPointHolder(GetGUIDLow());
-    }
-
-    if (m_comboPoints > 5)
-    {
-        m_comboPoints = 5;
-    }
-    if (m_comboPoints < 0)
-    {
-        m_comboPoints = 0;
-    }
-
-    SendComboPoints();
-}
-
-/**
- * @brief Clears the player's combo points and current combo target.
- */
-void Player::ClearComboPoints()
-{
-    if (!m_comboTargetGuid)
-    {
-        return;
-    }
-
-    // without combopoints lost (duration checked in aura)
-    RemoveSpellsCausingAura(SPELL_AURA_RETAIN_COMBO_POINTS);
-
-    m_comboPoints = 0;
-
-    SendComboPoints();
-
-    if (Unit* target = sObjectAccessor.GetUnit(*this, m_comboTargetGuid))
-    {
-        target->RemoveComboPointHolder(GetGUIDLow());
-    }
-
-    m_comboTargetGuid.Clear();
 }
 
 /**
